Backend/Easyocr/requirements.txt
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118


torchaudio                0.12.1               py38_cu113    pytorch
torchvision               0.13.1               py38_cu113    pytorch
pytorch                   1.12.1          py3.8_cuda11.3_cudnn8_0    pytorch
pytorch-mutex             1.0                        cuda    pytorch


pip install fire
pip install lmdb
pip install opencv-python
pip install natsort
pip install nltk